requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dj-database-url==2.3.0
Django==4.2.17
django-imagefield==0.21.1
django-versatileimagefield==3.1
djangorestframework==3.15.2
drf-yasg==1.21.8
flatbuffers==24.12.23
fonttools==4.55.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.4.1
mysqlclient==2.2.6
namex==0.0.8
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.10
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.13.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
wrapt==1.17.2
zipp==3.21.0
